+++ buildpacks/go/src/main.rs
@@ -12,6 +12,7 @@ mod tgz;
 use bullet_stream::global::print;
 use bullet_stream::style;
 use heroku_go_utils::vrs::GoVersion;
+use indoc::formatdoc;
 use layers::build::{handle_build_layer, BuildLayerError};
 use layers::deps::{handle_deps_layer, DepsLayerError};
 use layers::dist::{handle_dist_layer, DistLayerError};
@@ -27,7 +28,7 @@ use libcnb::{buildpack_main, Buildpack, Env};
 use libherokubuildpack::inventory::artifact::{Arch, Os};
 use libherokubuildpack::inventory::Inventory;
 use sha2::Sha256;
-use std::env::{self, consts};
+use std::env::consts;
 use std::path::Path;
 use std::time::Instant;
 
@@ -65,12 +66,23 @@ impl Buildpack for GoBuildpack {
         print::bullet("Reading build configuration");
 
         let started = Instant::now();
-        let mut go_env = Env::new();
-        env::vars()
-            .filter(|(k, _v)| k == "PATH")
-            .for_each(|(k, v)| {
-                go_env.insert(k, v);
-            });
+        let mut go_env = Env::from_current();
+        let go_prefixed_envs = go_env
+            .iter()
+            .map(|(key, _)| key.to_string_lossy())
+            .filter(|key| key.starts_with("GO"))
+            .map(bullet_stream::style::value)
+            .collect::<Vec<_>>()
+            .join(", ");
+
+        if !go_prefixed_envs.is_empty() {
+            print::warning(formatdoc! {"
+                WARNING: Found `GO` prefixed environment variables not set by this buildpack.
+                These variables may impact build or runtime behavior in unexpected ways.
+
+                Environment variables: {go_prefixed_envs}
+            "});
+        }
 
         let inv: Inventory<GoVersion, Sha256, Option<()>> =
             toml::from_str(INVENTORY).map_err(GoBuildpackError::InventoryParse)?;